feat: add lib and inputs args to overlay
Overlays previously would only receive a "channels" argument and the
rest of the attribute set would be the flake's inputs. This is now
changing to support passing "lib", the same instance you would get
in your system configuration, and "inputs", an attribute set of the
flake's inputs. Referencing flake inputs by name here is now
deprecated and will be removed in a future update.